LPD-93809 Separate token mapping UI in several react components for clarity
